pyproject.toml
[tool.poetry]
name = "kitchencompanion"
version = "0.1.0"
description = ""
authors = ["Kamil Wilczek <kamilwilczek@wp.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.4"
packaging = "^23.2"
python-decouple = "^3.8"
pytz = "^2023.3.post1"
sqlparse = "^0.4.4"
types-pytz = "^2023.3.1.1"
types-pyyaml = "^6.0.12.12"
tzdata = "^2023.3"
djoser = "^2.2.0"
django-axes = "^6.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
mypy = "^1.5.1"
isort = "^5.12.0"
freezegun = "^1.2.2"
pytest-django = "^4.5.2"
django-stubs = "^4.2.4"
django-stubs-ext = "^4.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
